create ActionCell
